Exercise 140: Play Bingo
In this exercise you will write a program that simulates a game of Bingo for a single
card. Begin by generating a list of all of the valid Bingo calls (B1 through O75). Once
the list has been created you can randomize the order of its elements by calling the
shufflefunction in therandommodule. Then your program should consume calls
out of the list, crossing out numbers on the card, until the card contains a crossed out
line (horizontal, vertical or diagonal). Simulate 1,000 games and report the minimum,
maximum and average number of calls that must be made before the card wins. Import
your solutions to Exercises 138 and 139 when completing this exercise.