Exercise 145: Find the Longest Word in a File
In this exercise you will create a Python program that identifies the longest word(s)
in a file. Your program should output an appropriate message that includes the length
of the longest word, along with all of the words of that length that occurred in the
file. Treat any group of non-white space characters as a word, even if it includes
numbers or punctuation marks.