﻿ Exercise 153: A Book with No “e”…
The novel “Gadsby” is over 50,000 words in length. While 50,000 words isn’t normally remarkable for a novel, it is in this case because none of the words in the book
use the letter “e”. This is particularly noteworthy when one considers that “e” is the
most common letter in English.
Write a program that reads a list of words from a file and determines what proportion of the words use each letter of the alphabet. Display the result for all 26
letters. Include an additional message identifying the letter that is used in the smallest proportion of the words. Your program should ignore any punctuation marks and
it should treat uppercase and lowercase letters as equivalent.