Exercise 155: Gender Neutral Names
Some names, like Ben and Jonathan, are normally only used for boys while names
like Rebbecca and Flora are normally only used for girls. Other names, like Chris
and Alex, may be used for both boys and girls.
Write a program that determines and displays all of the baby names that were
used for both boys and girls in a year specified by the user. Your program should
generate an appropriate message if there were no gender neutral names in the selected
year. Display an appropriate error message if you do not have data for the year
requested by the user. Additional details about the baby names data set are included
in Exercise 154.