Exercise 156: Most Births in a given Time Period
Write a program that uses the baby names data set described in Exercise 154 to
determine which names were used most often within a time period. Have the user
supply the first and last years of the range to analyze. Display the boy’s name and
the girl’s name given to the most children during the indicated years.